refactor: split onboard model selection
